feat: added the config file integration and an example config file
